+++ BlogMvcCore/Views/Post/ViewPostAndComments.cshtml
@@ -9,6 +9,8 @@
             <td>
                 <p>Let's start discussing...</p>
                 <p></p>
+                @*//TODO: change controller name to Comment
+                *@
                 @using (Html.BeginForm("AddComment", "User", FormMethod.Post))
                 {
                     @Html.TextBox("commentText", "", htmlAttributes: true)
@@ -18,7 +20,7 @@
                 <hr class="hr1">
                 @if (Model.Comments.Count() != 0)
                 {
-            <tr><td>@await Html.PartialAsync("_PostComment", Model.Comments)</td> </tr>
+            <tr><td>@Html.Partial("~/Views/User/_PostComment.cshtml", Model.Comments)</td> </tr>
                 }
     </table>
 </div>
