+++ BlogMvcCore/Views/Post/_AddPost.cshtml
@@ -1,5 +1,5 @@
 ﻿@model User
-@using (Html.BeginForm("AddPost", "User", FormMethod.Post))
+@using (Html.BeginForm("AddPost", "Post", FormMethod.Post))
 {
     <table>
         <tr><td><h5><em>Create a new post:</em></h5></td></tr>
