--- BlogMvcCore/Views/User/ViewRecentAddedPostList.cshtml
@@ -1,42 +1,0 @@
-﻿@model List<Post>
-
-<div>
-    <table border="1">
-        <tr>
-                @if (Model.Count > 0)
-                {
-                    @foreach (var item in Model)
-                    {
-                        @if (item.Text.Count() > 500)
-                        {
-                    <tr>
-                        <td>
-                            <p></p>
-                                    @item.Title
-                            <p></p>
-                                    @item.Text.Substring(0,500) @Html.ActionLink("view more","ViewPostAndComments", "User", new {postID=item.ID}, new {target="_blank"})
-                            <p></p>
-                        </td>
-                    </tr>
-                        }
-                        else
-                        {
-                    <tr>
-                        <td>
-                            <p></p>
-                                    @item.Title
-                            <p></p>
-                                    @item.Text @Html.ActionLink("view more","ViewPostAndComments","User", new {postID=item.ID}, new {target="_blank"})
-                            <p></p>
-                        </td>
-                    </tr>
-                        }
-                    }
-                }
-                else
-                {
-            <p>Here's nothing to show you. Write a post at first...</p>
-                }
-        </tr>
-    </table>
-</div>
